Consolidated Useful & Useless Code; Migrated Data Moved the scripts from distance_data_dump to FlavinDB and moved the data into Data repository
This commit is mostly cosmetic and administrative. The most recent data was
all generated using these scripts because the :Filter: class was difficult to
use and mildly broken for some reason. The goal here is to make the data
generation process a more transparent/coherent and easy to contribute to.

TODOs Left:
    - Add pydoc support for the code, objects and executable
    - Finish README.md with how to use the script and tool
    - Add some improvement goals for the script
    - Add URLs to get to Data